requirements.txt
Django>=5.0
djangorestframework>=3.14
psycopg2-binary>=2.9
drf-spectacular>=0.27
django-filter>=23.5
python-dotenv>=1.0